+++ express-zod-api/src/index.ts
@@ -28,6 +28,7 @@ export {
 } from "./errors";
 export { testEndpoint, testMiddleware } from "./testing";
 export { EventStreamFactory } from "./sse";
+export type { EventStreamFactoryOptions } from "./sse";
 
 export { ez } from "./proprietary-schemas";